+++ src/Factors/src/PreintegratedCentroidalDynamicsMeasurements.cpp
@@ -86,6 +86,7 @@ constexpr double massThreshold{1e-6};
 #define A_phi_phi(A)  (A)->block<3,3>(0,0)
 #define A_cdot_phi(A) (A)->block<3,3>(3,0)
 #define A_c_phi(A)    (A)->block<3,3>(6,0)
+#define A_c_cdot(A)   (A)->block<3,3>(6,3)
 #define A_ha_phi(A)   (A)->block<3,3>(9,0)
 
 
@@ -228,7 +229,7 @@ bool PreintegratedCDMCumulativeBias::updateCDMModelComputations(const std::share
     m_currModelComp.contactFrameLeverArmsInBase.clear();
     m_currModelComp.contactFrameRotationsInBase.clear();
     m_currModelComp.netExtForceInBase.setZero();
-    m_currModelComp.netExtForceInBase.setZero();
+    m_currModelComp.netExtTorqueInBase.setZero();
     m_currModelComp.netUnbiasedExtForceInBase.setZero();
     m_currModelComp.netUnbiasedExtTorqueInBase.setZero();
 
@@ -285,9 +286,9 @@ void PreintegratedCDMCumulativeBias::updateCDMGyroComputations(const gtsam::Vect
 {
     // DRkk+1  = DRik.T DRik+1 = Exp(omega dt)
     m_currGyroComp.unbiasedGyro = m_imuBiasHat.correctGyroscope(localGyroMeas);
-
     const gtsam::Vector3& omegaBar = m_currGyroComp.unbiasedGyro;
     m_currGyroComp.DRkkplusone = gtsam::Rot3::Expmap(omegaBar*dt);
+
     m_currGyroComp.Jr = gtsam::so3::DexpFunctor(omegaBar*dt).dexp();
     m_currGyroComp.DRik = m_deltaRij;
 }
@@ -317,7 +318,8 @@ void PreintegratedCDMCumulativeBias::prepareA(const CDMModelComputations& modelC
 
     A_phi_phi(&m_A) = DRkkplusone.inverse().matrix();
     A_cdot_phi(&m_A)= -(1/m)*DRik*Sf*dt;
-    A_c_phi(&m_A) = -(3/2*m)*DRik*Sf*dtSq;
+    A_c_phi(&m_A) = -(0.5/m)*DRik*Sf*dtSq;
+    A_c_cdot(&m_A) = gtsam::I_3x3*dt;
     A_ha_phi(&m_A) = -DRik*Stau*dt;
 }
 
@@ -325,18 +327,21 @@ void PreintegratedCDMCumulativeBias::prepareG(const CDMModelComputations& modelC
                                               const CDMGyroComputations& gyroComp,
                                               const double& dt)
 {
+    // obsDim = dim(gyro, com position, external wrenches)
     const std::size_t obsDim = 6 + modelComp.nrExtWrenches*6;
+    // obsWBiasDim = obsDim + dim(gyroBias, cdmBias)
+    const std::size_t obsWBiasDim = obsDim + 3 + gtsam::CDMBiasCumulative::dimension;
     const auto& m = modelComp.mass;
     const auto oneByM = (1/m);
-    const auto threeByTwoM{3/2*m};
+    const auto threeByTwoM{1.5/m};
     const auto& DRik = gyroComp.DRik.matrix();
 
     // both m_B and m_Sigma_n dimensions
     // might change every step depending
     // on number of external wrenches
-    m_B.resize(m_residualDim, obsDim);
+    m_B.resize(m_residualDim, obsWBiasDim);
     m_B.setZero();
-    m_Sigma_n.resize(obsDim, obsDim);
+    m_Sigma_n.resize(obsWBiasDim, obsWBiasDim);
     m_Sigma_n.setZero();
 
     const std::size_t comIdxOffset = 3 + modelComp.nrExtWrenches*6;
@@ -378,6 +383,8 @@ void PreintegratedCDMCumulativeBias::prepareG(const CDMModelComputations& modelC
     m_Sigma_n.block<3,3>(bfOffset, bfOffset) = m_p->getContactForceBiasCovariance();
     m_Sigma_n.block<3,3>(btOffset, btOffset) = m_p->getContactTorqueBiasCovariance();
 
+    m_B.block<12, 12>(12, bgOffset).setIdentity();
+
     m_G = m_B*m_Sigma_n*(m_B.transpose());
 }
 
@@ -403,6 +410,7 @@ bool PreintegratedCDMCumulativeBias::update(const std::shared_ptr<iDynTree::KinD
         return false;
     }
 
+    m_deltaTij += dt;
     updatePreintegratedMeasurements(m_currModelComp, m_currGyroComp, dt);
     propagatePreintegrationCovariance(m_currModelComp, m_currGyroComp, dt);
     updateBiasJacobians(m_currModelComp, m_currGyroComp, dt);
@@ -419,10 +427,11 @@ void PreintegratedCDMCumulativeBias::updatePreintegratedMeasurements(const CDMMo
     const auto& ExpOmegaDt = gyroComp.DRkkplusone;
     const gtsam::Vector3& B_f_net = modelComp.netUnbiasedExtForceInBase;
     const gtsam::Vector3& B_tau_net = modelComp.netUnbiasedExtTorqueInBase;
+    const gtsam::Vector3 DRikfdt = DRik.rotate(B_f_net)*dt;
 
     m_deltaRij = m_deltaRij*ExpOmegaDt;
-    m_deltaCdotij += (1/m)*DRik.rotate(B_f_net)*dt;
-    m_deltaCij += (3/2*m)*DRik.rotate(B_f_net)*dt*dt;
+    m_deltaCdotij += (1/m)*DRikfdt;
+    m_deltaCij += m_deltaCdotij*dt + (0.5/m)*DRikfdt*dt;
     m_deltaHaij += DRik.rotate(B_tau_net)*dt;
 }
 
@@ -433,7 +442,7 @@ void PreintegratedCDMCumulativeBias::updateBiasJacobians(const CDMModelComputati
     double dtSq{dt*dt};
     const auto& m = modelComp.mass;
     const double oneByM = (1/m);
-    const double threeByTwoM = (3/2*m);
+    const double halfByM = (0.5/m);
     const gtsam::Matrix3& DRik = gyroComp.DRik.matrix();
 
     const auto& DRkkplusone = gyroComp.DRkkplusone;
@@ -451,8 +460,8 @@ void PreintegratedCDMCumulativeBias::updateBiasJacobians(const CDMModelComputati
     m_delRdelBiasGyro = DRkkplusoneT*dRdbg_old - gyroComp.Jr*dt;
     m_delCdotdelBiasGyro += (-oneByM*DRik_Sf_dRbdbg*dt);
     m_delCdotdelBiasNetForce += (-oneByM*DRik_dt);
-    m_delCdelBiasGyro += (-threeByTwoM*DRik_Sf_dRbdbg*dtSq);
-    m_delCdelBiasNetForce += (-threeByTwoM*DRik_dt*dt);
+    m_delCdelBiasGyro += (-halfByM*DRik_Sf_dRbdbg*dtSq + m_delCdotdelBiasGyro*dt);
+    m_delCdelBiasNetForce += (-halfByM*DRik_dt*dt + m_delCdotdelBiasNetForce*dt);
     m_delHadelBiasGyro += (-DRik_Stau_dRbdbg*dt);
     m_delHadelBiasNetTorque += (-DRik_dt);
     m_delHadelBiasCOMPosition += (-DRik_Sf*dt);
@@ -538,9 +547,9 @@ gtsam::Vector PreintegratedCDMCumulativeBias::computeErrorAndJacobians(
         H2->resize(m_residualDim, 3);
         H2->setZero();
         // H22 = dr_DCdotij/dcdot_i
-        J_rDcdotij_cdot_i(H2) = -gtsam::I_3x3;
+        J_rDcdotij_cdot_i(H2) = -R_i_T.matrix();
         // H32 = dr_DCij/dcdot_i
-        J_rDcij_cdot_i(H2) = -gtsam::I_3x3*m_deltaTij;
+        J_rDcij_cdot_i(H2) = -R_i_T.matrix()*m_deltaTij;
     }
 
     // dr/dc_i
@@ -549,7 +558,7 @@ gtsam::Vector PreintegratedCDMCumulativeBias::computeErrorAndJacobians(
         H3->resize(m_residualDim, 3);
         H3->setZero();
         // H33 = dr_DCij/dc_i
-        J_rDcij_c_i(H3) = -gtsam::I_3x3;
+        J_rDcij_c_i(H3) = -R_i_T.matrix();
     }
 
     // dr/dha_i
@@ -576,7 +585,7 @@ gtsam::Vector PreintegratedCDMCumulativeBias::computeErrorAndJacobians(
         H6->resize(m_residualDim, 3);
         H6->setZero();
         // H26 = dr_DCdotij/dcdot_j
-        J_rDcdotij_cdot_j(H6) = R_i_T_R_j.matrix();
+        J_rDcdotij_cdot_j(H6) = R_i_T.matrix();
     }
 
     // dr/dc_j
@@ -585,7 +594,7 @@ gtsam::Vector PreintegratedCDMCumulativeBias::computeErrorAndJacobians(
         H7->resize(m_residualDim, 3);
         H7->setZero();
         // H37 = dr_DCij/dc_j
-        J_rDcij_c_j(H7) = R_i_T_R_j.matrix();
+        J_rDcij_c_j(H7) = R_i_T.matrix();
     }
 
     // dr/dha_j